refactor: add relation between post entity and user entity refactor: add util object for getting request user

+++ src/main/java/com/example/spring_basic/post/adapter/out/persistence/repository/PostRepositoryImpl.java
@@ -1,11 +1,13 @@
 package com.example.spring_basic.post.adapter.out.persistence.repository;
 
-import com.example.spring_basic.global.error.exception.NotFoundException;
+import com.example.spring_basic.global.error.exception.forbidden.ForbiddenException;
+import com.example.spring_basic.global.error.exception.notfound.NotFoundException;
 import com.example.spring_basic.post.adapter.out.persistence.mapper.JpaPostMapper;
 import com.example.spring_basic.post.adapter.out.persistence.model.PostModel;
 import com.example.spring_basic.post.domain.entity.Post;
 import com.example.spring_basic.post.domain.port.in.dto.PostCreationInput;
 import com.example.spring_basic.post.domain.port.out.repository.PostRepository;
+import com.example.spring_basic.user.domain.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -21,8 +23,9 @@ public class PostRepositoryImpl implements PostRepository {
     private final JpaPostMapper mapper;
 
     @Override
-    public void save(PostCreationInput input) {
-        jpaPostRepository.save(mapper.mapToJpaModelFromCreationInput(input));
+    public Post save(PostCreationInput input, User user) {
+        PostModel newPost = jpaPostRepository.save(mapper.mapToJpaModelFromCreationInput(input, user));
+        return mapper.mapToDomainEntity(newPost);
     }
 
     @Override
@@ -39,10 +42,19 @@ public class PostRepositoryImpl implements PostRepository {
     }
 
     @Override
-    public Post update(Post input) {
+    public Post update(Post input, Long requestUserId) {
         return jpaPostRepository.findById(input.getId())
+                .map(model -> checkOwnership(model, requestUserId))
                 .map(model -> updatePost(input, model))
                 .orElseThrow(() -> new NotFoundException("Post not found with id: " + input.getId()));
+    }
+
+    private PostModel checkOwnership(PostModel model, Long requestUserId) {
+        if (!model.getUser().getId().equals(requestUserId)) {
+            throw new ForbiddenException("Request user is not the owner of this post.");
+        }
+        return model;
+
     }
 
     private Post updatePost(Post input, PostModel model) {
@@ -58,7 +70,8 @@ public class PostRepositoryImpl implements PostRepository {
     }
 
     @Override
-    public void delete(Long id) {
+    public void delete(Long id, Long requestUserId) {
+        checkOwnership(jpaPostRepository.getReferenceById(id), requestUserId);
         jpaPostRepository.deleteById(id);
     }
 }
